+++ src/Data/TypedEncoding/Instances/Support/Helpers.hs
@@ -39,13 +39,6 @@ foldEnc :: forall (xs2 :: [Symbol]) (xs1 :: [Symbol]) f c s1 s2
            => c -> (s1 -> s2 -> s2) -> s2 -> f (Enc xs1 c s1) -> Enc xs2 c s2
 foldEnc c f sinit = unsafeSetPayload c . foldr f sinit . fmap getPayload 
 
--- | Similar to 'foldEnc', assumes that destination payload has @IsString@ instance and uses @""@ as base case. 
---
--- @since 0.2.0.0
-foldEncStr :: forall (xs2 :: [Symbol]) (xs1 :: [Symbol]) f c s1 s2 
-             . (Foldable f, Functor f, IsString s2) 
-             => c -> (s1 -> s2 -> s2) -> f (Enc xs1 c s1) -> Enc xs2 c s2
-foldEncStr c f = foldEnc c f ""
 
 -- | Similar to 'foldEnc', works with untyped 'CheckedEnc'
 --
@@ -55,16 +48,11 @@ foldCheckedEnc :: forall (xs2 :: [Symbol]) f c s1 s2
              => c -> ([EncAnn] -> s1 -> s2 -> s2) -> s2 -> f (CheckedEnc c s1) -> Enc xs2 c s2
 foldCheckedEnc c f sinit = unsafeSetPayload c . foldr (uncurry f) sinit . fmap getCheckedEncPayload
 
--- | Similar to 'foldEncStr', works with untyped 'CheckedEnc'
---
--- @since 0.2.0.0
-foldCheckedEncStr :: forall (xs2 :: [Symbol]) f c s1 s2 . (Foldable f, Functor f, IsString s2) => c -> ([EncAnn] -> s1 -> s2 -> s2) -> f (CheckedEnc c s1) -> Enc xs2 c s2
-foldCheckedEncStr c f  = foldCheckedEnc c f ""
 
 
 -- * Composite encoding: Recreate and Encode helpers
 
--- | Splits composite payload into homogenious chunks
+-- | Splits composite payload into homogeneous chunks
 --
 -- @since 0.2.0.0
 splitPayload :: forall (xs2 :: [Symbol]) (xs1 :: [Symbol]) c s1 s2 . 
@@ -75,12 +63,13 @@ splitPayload f (UnsafeMkEnc _ c s1) = map (UnsafeMkEnc Proxy c) (f s1)
    
 -- | Untyped version of 'splitPayload'
 --
--- @since 0.2.0.0
-splitSomePayload :: forall c s1 s2 . 
+-- (renamed from @splitCheckedPayload@ in previous versions)
+-- @since 0.5.0.0
+splitCheckedPayload :: forall c s1 s2 . 
              ([EncAnn] -> s1 -> [([EncAnn], s2)]) 
              -> CheckedEnc c s1 
              -> [CheckedEnc c s2]
-splitSomePayload f (UnsafeMkCheckedEnc ann1 c s1) = map (\(ann2, s2) -> UnsafeMkCheckedEnc ann2 c s2) (f ann1 s1)
+splitCheckedPayload f (UnsafeMkCheckedEnc ann1 c s1) = map (\(ann2, s2) -> UnsafeMkCheckedEnc ann2 c s2) (f ann1 s1)
 
 
 -- * Utility combinators