+++ ChangeLog.md
@@ -8,10 +8,12 @@
 - More module renaming to separate internal implementation code and code targeting examples
 - (post 0.5) "enc-B64" will be moved to a different package (more distant goal)
 
-## 0.5.0
+## Unreleased 0.5.0
 
 - Data.TypedEncoding.Instances.Do.Sample moved to Examples
 - Examples.TypedEncoding folder re-org
+- `Data.TypedEncoding.Instances.Support.Helpers` removed `foldEncStr`, `foldCheckedEncStr` 
+   renamed  `splitSomePayload` to `splitCheckedPayload`
 
 
 ## 0.4.2